# Fixes performance bottleneck in subset_map/extend