--- a/Makefile.conf
+++ b/Makefile.conf
@@ -43,6 +43,7 @@ SOURCES = \
 	to_c.cpp \
 	to_value.cpp \
 	source_map.cpp \
+	subset_map.cpp \
 	error_handling.cpp \
 	memory/SharedPtr.cpp \
 	utf8_string.cpp \