--- a/src/context.hpp
+++ b/src/context.hpp
@@ -48,7 +48,7 @@ namespace Sass {
     std::vector<char*> strings;
     std::vector<Resource> resources;
     std::map<const std::string, StyleSheet> sheets;
-    Subset_Map<std::string, std::pair<Complex_Selector_Obj, Compound_Selector_Obj> > subset_map;
+    Subset_Map subset_map;
     std::vector<Sass_Import_Entry> import_stack;
 
     struct Sass_Compiler* c_compiler;