You can expect to see a 2x performance increase in bigger sass code bases. With more extend usage it can go as high as 10x performance increase.

Test with styles of a real corporate website showed a performance increase of 3x (from 750ms to 250ms). When I imported the main stylesheet 10x times, compile time went from 100s to 5s (~20x increase).

This indecently (according to probe-todo) also fixes https://github.com/sass/libsass/issues/2056 (false positive?)?