@xzyfer 2016-12-19
It's noting that the expected behaviour of #2056 has changed. The bug is now incorrect, see the final comments.