@xzyfer 2016-12-19
Nice one. This approach is fine by me. It's worth nothing that the extends behaviour has changed slightly in dart-sass, whose behaviour is going to be the reference implementation.