@mgreter 2016-12-19
As stated, this looks like a false positive (therefore meaning probe-todo is still broken) ...