@xzyfer 2016-12-19
Yeah. Just wanted to make sure you were aware of the difference between Ruby Sass and Dart Sass since you're working on the extends code.