@mgreter 2016-12-20
FWIW: made some performance assesments accross different implementations:

```scss
.clearfix {
  &:after {
    content: "";
    display: table;
    clear: both;
  }
}

@for $n from 1 through 100 {
	@for $i from 1 through 100 {
		.foo#{$i} { @extend .clearfix; }
	}
}
```

```cmd
$ bin/sassc.exe -v
sassc: 3.4.1
libsass: 3.4.0-19-g414b
sass2scss: 1.1.0
sass: 3.4
$ time sassc test.scss
real    0m0.279s
user    0m0.000s
sys     0m0.015s
```

```cmd
$ sass -v
Sass 3.4.23 (Selective Steve)
$ time sass test.scss
real    0m2.560s
user    0m1.984s
sys     0m0.561s
```

```cmd
$ dart-sass --version
1.0.0-alpha.6
$ time dart-sass test.scss
real    0m45.833s
user    0m0.000s
sys     0m0.000s
```

//CC @xzyfer @nex3